requirements.txt
annotated-types==0.7.0
attrs==22.2.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coverage==7.8.0
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
httplib2==0.22.0
idna==3.10
iniconfig==2.0.0
mypy_extensions==1.1.0
numpy==1.26.2
packaging==23.0
pathspec==0.12.1
pdfkit==1.0.0
platformdirs==4.3.8
pluggy==1.0.0
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
PyJWT==2.10.1
pyparsing==3.2.3
pytest==7.2.1
pytest-cov==6.1.1
python-dotenv==1.1.0
requests==2.32.3
requests-mock==1.12.1
rsa==4.9.1
typing-inspection==0.4.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
